requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
astunparse==1.6.3
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h5py==3.13.0
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.0.9
networkx==3.5
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pfzy==0.3.4
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.1.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2